# [Gekidou] Performance & UI fixes