--- a/app/actions/remote/entry/common.ts
+++ b/app/actions/remote/entry/common.ts
@@ -83,8 +83,8 @@ export const fetchAppEntryData = async (serverUrl: string, since: number, initia
         // If no initial team was set in the database but got teams in the response
         const config = await getConfig(database);
         const teamOrderPreference = getPreferenceValue(prefData.preferences || [], Preferences.TEAMS_ORDER, '', '') as string;
-        const teamMembers = teamData.memberships.filter((m) => m.delete_at === 0).map((m) => m.team_id);
-        const myTeams = teamData.teams!.filter((t) => teamMembers?.includes(t.id));
+        const teamMembers = new Set(teamData.memberships.filter((m) => m.delete_at === 0).map((m) => m.team_id));
+        const myTeams = teamData.teams!.filter((t) => teamMembers.has(t.id));
         const defaultTeam = selectDefaultTeam(myTeams, meData.user?.locale || DEFAULT_LOCALE, teamOrderPreference, config?.ExperimentalPrimaryTeam);
         if (defaultTeam?.id) {
             chData = await fetchMyChannelsForTeam(serverUrl, defaultTeam.id, includeDeletedChannels, since, fetchOnly);
@@ -136,11 +136,11 @@ export const fetchAppEntryData = async (serverUrl: string, since: number, initia
 
     if (data.chData?.channels) {
         const removeChannelIds: string[] = [];
-        const fetchedChannelIds = data.chData.channels.map((channel) => channel.id);
+        const fetchedChannelIds = new Set(data.chData.channels.map((channel) => channel.id));
 
         const channels = await queryAllChannelsForTeam(database, initialTeamId).fetch();
         for (const channel of channels) {
-            if (!fetchedChannelIds.includes(channel.id)) {
+            if (!fetchedChannelIds.has(channel.id)) {
                 removeChannelIds.push(channel.id);
             }
         }