--- a/app/actions/remote/user.ts
+++ b/app/actions/remote/user.ts
@@ -14,7 +14,7 @@ import {General} from '@constants';
 import DatabaseManager from '@database/manager';
 import {debounce} from '@helpers/api/general';
 import NetworkManager from '@init/network_manager';
-import {queryChannelsByTypes} from '@queries/servers/channel';
+import {getMembersCountByChannelsId, queryChannelsByTypes} from '@queries/servers/channel';
 import {getCurrentTeamId, getCurrentUserId} from '@queries/servers/system';
 import {getCurrentUser, getUserById, prepareUsers, queryAllUsers, queryUsersById, queryUsersByUsername} from '@queries/servers/user';
 import {removeUserFromList} from '@utils/user';
@@ -116,9 +116,21 @@ export const fetchProfilesInChannel = async (serverUrl: string, channelId: strin
 
 export const fetchProfilesPerChannels = async (serverUrl: string, channelIds: string[], excludeUserId?: string, fetchOnly = false): Promise<ProfilesPerChannelRequest> => {
     try {
+        const operator = DatabaseManager.serverDatabases[serverUrl]?.operator;
+        if (!operator) {
+            return {error: `${serverUrl} database not found`};
+        }
+
+        const {database} = operator;
+
+        // let's filter those channels that we already have the users
+        const membersCount = await getMembersCountByChannelsId(database, channelIds);
+        const channelsToFetch = channelIds.filter((c) => !membersCount[c]);
+
         // Batch fetching profiles per channel by chunks of 50
-        const channels = chunk(channelIds, 50);
+        const channels = chunk(channelsToFetch, 50);
         const data: ProfilesInChannelRequest[] = [];
+
         for await (const cIds of channels) {
             const requests = cIds.map((id) => fetchProfilesInChannel(serverUrl, id, excludeUserId, true));
             const response = await Promise.all(requests);
@@ -126,32 +138,29 @@ export const fetchProfilesPerChannels = async (serverUrl: string, channelIds: st
         }
 
         if (!fetchOnly) {
-            const operator = DatabaseManager.serverDatabases[serverUrl]?.operator;
-            if (operator) {
-                const modelPromises: Array<Promise<Model[]>> = [];
-                const users = new Set<UserProfile>();
-                const memberships: Array<{channel_id: string; user_id: string}> = [];
-                for (const item of data) {
-                    if (item.users?.length) {
-                        item.users.forEach((u) => {
-                            users.add(u);
-                            memberships.push({channel_id: item.channelId, user_id: u.id});
-                        });
-                    }
+            const modelPromises: Array<Promise<Model[]>> = [];
+            const users = new Set<UserProfile>();
+            const memberships: Array<{channel_id: string; user_id: string}> = [];
+            for (const item of data) {
+                if (item.users?.length) {
+                    item.users.forEach((u) => {
+                        users.add(u);
+                        memberships.push({channel_id: item.channelId, user_id: u.id});
+                    });
                 }
-                modelPromises.push(operator.handleChannelMembership({
-                    channelMemberships: memberships,
-                    prepareRecordsOnly: true,
-                }));
-                const prepare = prepareUsers(operator, Array.from(users).filter((u) => u.id !== excludeUserId));
-                if (prepare) {
-                    modelPromises.push(prepare);
-                }
+            }
+            modelPromises.push(operator.handleChannelMembership({
+                channelMemberships: memberships,
+                prepareRecordsOnly: true,
+            }));
+            const prepare = prepareUsers(operator, Array.from(users).filter((u) => u.id !== excludeUserId));
+            if (prepare) {
+                modelPromises.push(prepare);
+            }
 
-                if (modelPromises.length) {
-                    const models = await Promise.all(modelPromises);
-                    await operator.batchRecords(models.flat());
-                }
+            if (modelPromises.length) {
+                const models = await Promise.all(modelPromises);
+                await operator.batchRecords(models.flat());
             }
         }
 
@@ -162,9 +171,8 @@ export const fetchProfilesPerChannels = async (serverUrl: string, channelIds: st
 };
 
 export const updateMe = async (serverUrl: string, user: Partial<UserProfile>) => {
-    const database = DatabaseManager.serverDatabases[serverUrl]?.database;
     const operator = DatabaseManager.serverDatabases[serverUrl]?.operator;
-    if (!database) {
+    if (!operator) {
         return {error: `${serverUrl} database not found`};
     }