--- a/app/actions/remote/user.ts
+++ b/app/actions/remote/user.ts
@@ -233,8 +233,13 @@ export const fetchStatusByIds = async (serverUrl: string, userIds: string[], fet
             const {database, operator} = DatabaseManager.serverDatabases[serverUrl];
             if (operator) {
                 const users = await queryUsersById(database, userIds).fetch();
+                const userStatuses = statuses.reduce((result, s) => {
+                    result[s.user_id] = s;
+                    return result;
+                }, {} as Record<string, UserStatus>);
+
                 for (const user of users) {
-                    const status = statuses.find((s) => s.user_id === user.id);
+                    const status = userStatuses[user.id];
                     user.prepareStatus(status?.status || General.OFFLINE);
                 }
 
@@ -271,7 +276,11 @@ export const fetchUsersByIds = async (serverUrl: string, userIds: string[], fetc
         if (userIds.includes(currentUser!.id)) {
             existingUsers.push(currentUser!);
         }
-        const usersToLoad = new Set(userIds.filter((id) => (!existingUsers.find((u) => u.id === id))));
+        const exisitingUsersMap = existingUsers.reduce((result, u) => {
+            result[u.id] = u;
+            return result;
+        }, {} as Record<string, UserModel>);
+        const usersToLoad = new Set(userIds.filter((id) => !exisitingUsersMap[id]));
         if (usersToLoad.size === 0) {
             return {users: [], existingUsers};
         }
@@ -308,8 +317,12 @@ export const fetchUsersByUsernames = async (serverUrl: string, usernames: string
 
     try {
         const currentUser = await getCurrentUser(operator.database);
-        const exisingUsers = await queryUsersByUsername(operator.database, usernames).fetch();
-        const usersToLoad = usernames.filter((username) => (username !== currentUser?.username && !exisingUsers.find((u) => u.username === username)));
+        const existingUsers = await queryUsersByUsername(operator.database, usernames).fetch();
+        const exisitingUsersMap = existingUsers.reduce((result, u) => {
+            result[u.username] = u;
+            return result;
+        }, {} as Record<string, UserModel>);
+        const usersToLoad = usernames.filter((username) => (username !== currentUser?.username && !exisitingUsersMap[username]));
         const users = await client.getProfilesByUsernames([...new Set(usersToLoad)]);
 
         if (!fetchOnly) {