#### Summary
This is a large PR, sorry for that.. we could split into smaller PR's but my idea was to use this to build and app that can be used to review a few things together.

To ease review, you could go by commits, the larger and most important part of the PR is around the database changes to improve performance, specially around `processRecords`.

Other important fixes around perf:
* Replace `indexOf` `includes` `find` and `findIndex` inside loops (aka for, filter, reduce, etc..) as one loop is will be O(n) and using any of this functions O(n) inside the loop will cause this to be O(n2). By changing it to a reduce to create a map or by using Set instead we'll keep the iteration as O(n)

Here are some numbers that I was able to get:
Records being processed: **7295** (this is the models involved in initial load with my account on community)

#### Prepare models
||Before | After |
|------|----------|---------|
| Run 1 | 9266ms | 3872ms |
| Run 2 | 11343ms | 3236ms|
| Run 3 | 9176ms | 3895ms |
| Run 4 | 9170ms | 3376ms |
| Run 5 | 10829ms | 3372ms |

#### Batch models
||Before | After |
|------|----------|---------|
| Run 1 | 2890ms | 2556ms |
| Run 2 | 2709ms | 1792ms|
| Run 3 | 1832ms | 1787ms |
| Run 4 | 1852ms | 1745ms |
| Run 5 | 2564ms | 1767ms |

#### Ticket Link
https://mattermost.atlassian.net/browse/MM-42986
https://mattermost.atlassian.net/browse/MM-42985
https://mattermost.atlassian.net/browse/MM-42984
https://mattermost.atlassian.net/browse/MM-42983
https://mattermost.atlassian.net/browse/MM-42982
https://mattermost.atlassian.net/browse/MM-42976

#### Checklist
<!--
Place an '[x]' (no spaces) in all applicable fields. Please remove unrelated fields.
-->
- [x] Added or updated unit tests (required for all new features)
- [x] Has UI changes
- [x] Includes text changes and localization file updates
- [x] Have tested against the 5 core themes to ensure consistency between them.

#### Device Information
This PR was tested on: Android 12 Pixel 4, iOS 15 iPhone 13 Pro Max, iPad 15 iPad Pro 11"

```release-note
NONE
```