@mattermod 2022-04-02
Successfully building:  
https://circleci.com/gh/mattermost/mattermost-mobile/21436  
https://circleci.com/gh/mattermost/mattermost-mobile/21435