@mattermod 2022-04-04
Successfully building:  
https://circleci.com/gh/mattermost/mattermost-mobile/21479  
https://circleci.com/gh/mattermost/mattermost-mobile/21478